requirements.txt
flask
requests
python-dotenv
asgiref
uvicorn